Refactor #49: Implemented Maven Changed directory structure and created pom.xml to work with maven
