+++ 1-DECIDE/src/test/Java/decideTests.java
@@ -1,10 +1,8 @@
-package test;
-import src.CMV;
-import src.PUM;
-import src.FUV;
-import src.MissileSystem;
-import src.Parameters;
-import src.Point;
+package src.test.Java;
+
+import src.Main.Java.MissileSystem;
+import src.Main.Java.Parameters;
+import src.Main.Java.Point;
 import java.util.Random;
 
 import org.junit.Test;
